% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLandmarkDataset.R
\name{createLandmarkDataset}
\alias{createLandmarkDataset}
\title{createLandmarkDataset}
\usage{
createLandmarkDataset(landmarkRepertory)
}
\arguments{
\item{landmarkRepertory}{The directory containing the landmark files.}
}
\value{
A combined data frame representing the landmark dataset.
}
\description{
Create a landmark dataset by combining flow cytometry data from multiple files.
}
\examples{
landmarkData <- createLandmarkDataset("/path/to/landmark/files")

}
