% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainAnnotationStatistics.R
\name{obtainAnnotationStatistics}
\alias{obtainAnnotationStatistics}
\title{obtainAnnotationStatistics}
\usage{
obtainAnnotationStatistics(enrichedFlowFrame, enrichColumnName)
}
\arguments{
\item{enrichedFlowFrame}{A FlowFrame object that has been enriched with annotation information.}

\item{enrichColumnName}{The name of the column containing the enrichment information.}
}
\value{
A data frame containing the population IDs, counts, and percentages.
}
\description{
Calculate annotation statistics from an enriched FlowFrame.
}
\examples{
stats <- obtainAnnotationStatistics(enrichedFlowFrame, "popIDXGBoost")

}
