% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCellAnnotation.R
\name{predictCellAnnotation}
\alias{predictCellAnnotation}
\title{predictCellAnnotation}
\usage{
predictCellAnnotation(model, flowFrame, enrichColumnName)
}
\arguments{
\item{model}{A trained predictive xgboost  model that will be used to make predictions on the flow cytometry data}

\item{flowFrame}{a flowFrame object containing the flow cytometry data to which the new annotations will be added.}

\item{enrichColumnName}{The name of the new column to be added to the flowFrame containing the predictions}
}
\value{
enrichedFlowFrame
}
\description{
The function predictCellAnnotation is designed to add predicted cell annotations to a flow cytometry dataset. This function takes a trained model, a flowFrame object, and a name for the new annotation column, then returns the flowFrame object with the added annotations.
}
