% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareUngatedData.R
\name{prepareUngatedData}
\alias{prepareUngatedData}
\title{prepareUngatedData}
\usage{
prepareUngatedData(flowFrame, markersInFCSFile, algo)
}
\arguments{
\item{flowFrame}{A flowFrame object containing flow cytometry data.}

\item{markersInFCSFile}{A vector of marker names present in the FCS file.}

\item{algo}{A trained model object (e.g., XGBoost) with feature names.}
}
\value{
An XGBoost DMatrix object with scaled data ready for prediction.
}
\description{
prepareUngatedData function is designed to preprocess flow cytometry data for prediction using an XGBoost model. This function extracts, from transformed fcs, and formats the relevant markers from the flow cytometry data to match the features used in the model, ultimately returning a data matrix ready for prediction.
}
\examples{
matrixForAnnotation <- prepareMatrix(fcs,  c("FSC-A","SSC-A","CD4","CD11b","CD44","CD5/Ly6G","IA/IE","CD317","CD62L","CD8a/SiglecF","F4/80","CD161","Ly6C","CD25","CD11c",'CD19'), xgboostModel)

}
