% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainXGBoostModel.R
\name{trainXGBoostModel}
\alias{trainXGBoostModel}
\title{trainXGBoostModel}
\usage{
trainXGBoostModel(referenceDataset, features, popIDColumn)
}
\arguments{
\item{referenceDataset}{A data frame containing the reference dataset with gated cell populations.}

\item{features}{A vector of markers to be used for training the model.}

\item{popIDColumn}{The name of the column in your dataset that contains the population IDs (popID).}
}
\value{
An XGBoost model object trained on the reference dataset.
}
\description{
Train an XGBoost model with a reference dataset that contains gated cell populations.
}
\examples{
trainXGBoostModel(referenceDataset, c("FSC-A", "SSC-A", "CD4", "CD11b", "CD44", "CD5/Ly6G", "IA/IE", "CD317", "CD62L", "CD8a/SiglecF", "F4/80", "CD161", "Ly6C", "CD25", "CD11c", "CD19"), "popID")

}
